user Bucket for rating limitation

+++ src/main/java/EventManager/EventManager/user/UserController.java
@@ -8,11 +8,16 @@ import EventManager.EventManager.event.beans.EventsSortByPopularityResults;
 import EventManager.EventManager.jpa.beans.Event;
 import EventManager.EventManager.jpa.beans.User;
 import EventManager.EventManager.remainder.EventRemainderService;
+import io.github.bucket4j.Bandwidth;
+import io.github.bucket4j.Bucket;
+import io.github.bucket4j.Refill;
 import jakarta.validation.Valid;
 import org.springframework.hateoas.EntityModel;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,13 +25,15 @@ import java.util.Optional;
 public class UserController {
     private final UserJpaService userJpaService;
     private final EventJpaService eventJpaService;
-
     private final EventRemainderService eventRemainderService;
+    private final Bucket bucket;
 
     public UserController(UserJpaService userJpaService, EventJpaService eventJpaService, EventRemainderService eventRemainderService) {
         this.userJpaService = userJpaService;
         this.eventJpaService = eventJpaService;
         this.eventRemainderService = eventRemainderService;
+        Bandwidth limit = Bandwidth.classic(20, Refill.greedy(20, Duration.ofMinutes(1)));
+        this.bucket = Bucket.builder().addLimit(limit).build();
     }
 
     @GetMapping(path = "/users/{id}")
@@ -37,7 +44,10 @@ public class UserController {
     @PostMapping(path = "/users/createUser")
     public ResponseEntity<User> createUser(@Valid @RequestBody User user){
         try {
-            return ResponseEntity.of(Optional.of(userJpaService.createUser(user)));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(Optional.of(userJpaService.createUser(user)));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -47,8 +57,11 @@ public class UserController {
     @DeleteMapping(path = "/users/{id}/deleteUser")
     public ResponseEntity<User> deleteUser(@PathVariable long id){
         try {
-            userJpaService.deleteUser(id);
-            return ResponseEntity.ok().build();
+            if (bucket.tryConsume(1)) {
+                userJpaService.deleteUser(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -57,8 +70,12 @@ public class UserController {
 
     @GetMapping(path = "/users/{id}/getUserEvents")
     public ResponseEntity<List<Event>> getEventsForUser(@PathVariable long id){
+
         try {
-        return ResponseEntity.of(Optional.of(eventJpaService.retrieveEventForUser(id)));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(Optional.of(eventJpaService.retrieveEventForUser(id)));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -68,7 +85,10 @@ public class UserController {
     @GetMapping(path = "/users/{id}/events/sortByCreationTime")
     public ResponseEntity<EventsSortByCreationDateResults> getEventsSortByCreationTimeForUser(@PathVariable long id){
         try {
-        return ResponseEntity.of(Optional.of(new EventsSortByCreationDateResults(eventJpaService.retrieveEventForUserSortByCreationTime(id))));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(Optional.of(new EventsSortByCreationDateResults(eventJpaService.retrieveEventForUserSortByCreationTime(id))));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -78,7 +98,10 @@ public class UserController {
     @GetMapping(path = "/users/{id}/events/sortByDate")
     public ResponseEntity<EventsSortByEventDateResults> getEventsSortByEventDateForUser(@PathVariable long id){
         try {
-        return ResponseEntity.of(Optional.of(new EventsSortByEventDateResults(eventJpaService.retrieveEventForUserSortByDate(id))));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(Optional.of(new EventsSortByEventDateResults(eventJpaService.retrieveEventForUserSortByDate(id))));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -87,7 +110,10 @@ public class UserController {
     @GetMapping(path = "/users/{id}/events/sortByPopularity")
     public ResponseEntity<EventsSortByPopularityResults> getEventsSortByPopularityForUser(@PathVariable long id){
         try {
-        return ResponseEntity.of(Optional.of(new EventsSortByPopularityResults(eventJpaService.retrieveEventForUserSortByPopularity(id))));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(Optional.of(new EventsSortByPopularityResults(eventJpaService.retrieveEventForUserSortByPopularity(id))));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -97,7 +123,10 @@ public class UserController {
     @GetMapping(path = "/users/{id}/events/locationFilter/{location}")
     public ResponseEntity<EventByLocationResults> getEventsByLocationForUser(@PathVariable long id, @PathVariable String location){
         try {
-        return ResponseEntity.of(Optional.of(new EventByLocationResults(eventJpaService.retrieveEventForUserByLocation(id,location))));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(Optional.of(new EventByLocationResults(eventJpaService.retrieveEventForUserByLocation(id, location))));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -107,9 +136,12 @@ public class UserController {
     @PostMapping(path = "/users/{id}/createEvent")
     public ResponseEntity<Event> createEvent(@PathVariable long id,@Valid @RequestBody Event event){
         try {
-            Event savedEvent = eventJpaService.createEventForUser(id,event);
-            eventRemainderService.runRemainder(id,savedEvent);
-            return ResponseEntity.of(Optional.of(savedEvent));
+            if (bucket.tryConsume(1)) {
+                Event savedEvent = eventJpaService.createEventForUser(id, event);
+                eventRemainderService.runRemainder(id, savedEvent);
+                return ResponseEntity.of(Optional.of(savedEvent));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }
         catch (Exception e){
            e.printStackTrace();
@@ -120,8 +152,10 @@ public class UserController {
     @GetMapping(path = "/users/{id}/getEventById/{eventId}")
     public ResponseEntity<Event> getEventById(@PathVariable long id,@PathVariable long eventId){
         try {
-
-            return ResponseEntity.of(eventJpaService.findEvent(eventId));
+            if (bucket.tryConsume(1)) {
+                return ResponseEntity.of(eventJpaService.findEvent(eventId));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }
         catch (Exception e){
             e.printStackTrace();
@@ -133,10 +167,12 @@ public class UserController {
     @DeleteMapping(path = "/users/{id}/deleteEvent")
     public ResponseEntity<User> deleteEvent(@RequestBody @Valid Event event){
         try {
-
-       eventJpaService.deleteEvent(event.getId());
-       eventRemainderService.deleteRemainder(event);
-       return ResponseEntity.ok().build();
+        if (bucket.tryConsume(1)) {
+            eventJpaService.deleteEvent(event.getId());
+            eventRemainderService.deleteRemainder(event);
+            return ResponseEntity.ok().build();
+        }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
             e.printStackTrace();
             return ResponseEntity.internalServerError().build();
@@ -146,9 +182,12 @@ public class UserController {
     @PostMapping(path = "/users/{id}/updateEvent")
     public ResponseEntity<Event> updateEvent(@PathVariable long id,@RequestBody @Valid Event event){
         try {
-           eventJpaService.updateEvent(event);
-           eventRemainderService.updateRemainder(id, event);
-           return ResponseEntity.of(Optional.of(event));
+            if (bucket.tryConsume(1)) {
+                eventJpaService.updateEvent(event);
+                eventRemainderService.updateRemainder(id, event);
+                return ResponseEntity.of(Optional.of(event));
+            }
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
         }catch (Exception e){
         e.printStackTrace();
         return ResponseEntity.internalServerError().build();
